move remaining file endpoints to TransfersController
